Allow the window to reset the size.Positions are represented by double.

+++ src/Common/DataStructure.h
@@ -106,18 +106,18 @@ public:
 
 class BaseShape {
 protected:
-    int posX, posY;
+    double posX, posY;
     int type;
     string name;
     double scaleX, scaleY, angle;
 public:
-    BaseShape(int posX, int posY, int type, const string &name, double scaleX, double scaleY, double angle) : posX(posX), posY(posY), type(type), name(name), scaleX(scaleX), scaleY(scaleY), angle(angle) {}
+    BaseShape(double posX, double posY, int type, const string &name, double scaleX, double scaleY, double angle) : posX(posX), posY(posY), type(type), name(name), scaleX(scaleX), scaleY(scaleY), angle(angle) {}
 
-    void setPosX(int posX) {
+    void setPosX(double posX) {
         BaseShape::posX = posX;
     }
 
-    void setPosY(int posY) {
+    void setPosY(double posY) {
         BaseShape::posY = posY;
     }
 
@@ -125,12 +125,12 @@ public:
         return type;
     }
 
-    int getPosX() const {
+    double getPosX() const {
         return posX;
     }
 
 
-    int getPosY() const {
+    double getPosY() const {
         return posY;
     }
 
@@ -170,10 +170,10 @@ public:
 class Line : public BaseShape {
 protected:
     Pen pen;
-    int x1, y1, x2, y2;
+    double x1, y1, x2, y2;
 public:
-    Line(int posX, int posY, int type, const string &name, double scaleX, double scaleY, double angle, const Pen &pen,
-          int x1, int y1, int x2, int y2) : BaseShape(posX, posY, type, name, scaleX, scaleY, angle),
+    Line(double posX, double posY, int type, const string &name, double scaleX, double scaleY, double angle, const Pen &pen,
+          double x1, double y1, double x2, double y2) : BaseShape(posX, posY, type, name, scaleX, scaleY, angle),
                                                                pen(pen),  x1(x1), y1(y1), x2(x2), y2(y2) {}
 
     const Pen &getPen() const {
@@ -184,35 +184,35 @@ public:
         Line::pen = pen;
     }
 
-    int getX1() const {
+    double getX1() const {
         return x1;
     }
 
-    void setX1(int x1) {
+    void setX1(double x1) {
         Line::x1 = x1;
     }
 
-    int getY1() const {
+    double getY1() const {
         return y1;
     }
 
-    void setY1(int y1) {
+    void setY1(double y1) {
         Line::y1 = y1;
     }
 
-    int getX2() const {
+    double getX2() const {
         return x2;
     }
 
-    void setX2(int x2) {
+    void setX2(double x2) {
         Line::x2 = x2;
     }
 
-    int getY2() const {
+    double getY2() const {
         return y2;
     }
 
-    void setY2(int y2) {
+    void setY2(double y2) {
         Line::y2 = y2;
     }
