+++ src/Model/Model.h
@@ -25,7 +25,7 @@ public:
     }
     void addLine(){
     }
-    void addLine(int centerX,int centerY,int x1,int y1,int x2,int y2);
+    void addLine(double centerX,double centerY,double x1,double y1,double x2,double y2);
     void SetPen(Pen pen){this->pen=pen;}
     const Pen* GetPen(){return &pen;}
     const Layouts* GetLayouts(){return &layouts;}
